atualizado

+++ funçoes/blibioteca.h
@@ -2,25 +2,25 @@
 #include <stdlib.h>
 
 float soma(float num1, float num2){
-	int resultado;
+	float resultado;
 	resultado= num1+num2;
 	return resultado;
 }
 
 float subtracao(float num1, float num2){
-	int resultado;
+	float resultado;
 	resultado = num1-num2;
 	return resultado;
 }
 
 float multiplicacao(float num1, float num2){
-	int resultado;
+	float resultado;
 	resultado = num1*num2;
 	return resultado;
 }
 
 float divisao(float num1, float num2){
-	int resultado;
+	float resultado;
 	resultado = num1/num2;
 	return resultado;
 }
